FormasGeometricas/src/main/java/org/mvpigs/formGeom/Cuadrado.java
package org.mvpigs.formGeom;

public class Cuadrado extends FiguraGeometrica {

    private double lado = 0d;

    public Cuadrado() {
        super();
    }

    public Cuadrado(double lado) {
        super();
        this.lado = lado;
    }

    public Cuadrado(String nombre, double lado) {
        super(nombre);
        this.lado = lado;
    }

    /**
     * @param lado the lado to set
     */
    public void setLado(double lado) {
        this.lado = lado;
    }

    /**
     * @return the lado
     */
    public double getLado() {
        return lado;
    }

    @Override
    public double calcularArea() {
        return lado * lado;
    }
}